+++ Source/InventoryPlugin/Public/Definitions.h
@@ -88,47 +88,3 @@ enum class EBagSlot : uint8
 	StagingArea = 21 UMETA(DisplayName = "StagingArea"),
 	BankPool = 22  UMETA(DisplayName = "BankingPool")
 };
-
-UENUM(BlueprintType)
-enum struct EItemSize : uint8
-{
-	Tiny UMETA(DisplayName = "Tiny"),
-	Small UMETA(DisplayName = "Small"),
-	Medium UMETA(DisplayName = "Medium"),
-	Large UMETA(DisplayName = "Large"),
-	Giant UMETA(DisplayName = "Giant")
-};
-
-USTRUCT(BlueprintType)
-struct INVENTORYPLUGIN_API FAttributeStruct
-{
-	GENERATED_BODY()
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Attributes")
-	int Str = 0;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Attributes")
-	int Sta = 0;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Attributes")
-	int Agi = 0;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Attributes")
-	int Dex = 0;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Attributes")
-	int Int = 0;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Attributes")
-	int Wis = 0;
-
-	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Attributes")
-	int Cha = 0;
-
-
-	FAttributeStruct& operator+=(const FAttributeStruct& OtherAttribute);
-
-	int Sum() const;
-
-	bool IsNotNull() const;
-};
